requirements.txt
requests
python-dotenv
gitpython
typing
tqdm
streamlit
pandas
matplotlib
seaborn